% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_tuning_results}
\alias{aggregate_tuning_results}
\title{Aggregate LORS results computed with different lambda, gamma}
\usage{
aggregate_tuning_results(tuning_results)
}
\arguments{
\item{D_list_validation}{}
}
\value{

}
\description{
Aggregate LORS results computed with different lambda, gamma
}
