% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_avg_correlation}
\alias{compute_avg_correlation}
\title{Compute average correlation across all responses}
\usage{
compute_avg_correlation(Y_list, X_list, indices_list, Beta)
}
\arguments{
\item{Beta}{}
}
\value{

}
\description{
Compute average correlation across all responses
}
