% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_avg_R2}
\alias{compute_avg_R2}
\title{Compute average R2 across all responses}
\usage{
compute_avg_R2(
  Y_list_test,
  X_list_test,
  indices_list_test,
  Y_list_train,
  indices_list_train,
  Beta
)
}
\arguments{
\item{Beta}{}
}
\value{

}
\description{
Compute average R2 across all responses
}
