% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_correlation}
\alias{compute_correlation}
\title{Compute correlation between actual and predicted per response}
\usage{
compute_correlation(Y_list, X_list, indices_list, Beta)
}
\arguments{
\item{Beta}{}
}
\value{

}
\description{
Compute correlation between actual and predicted per response
}
