% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{compute_R2}
\alias{compute_R2}
\title{Compute R2 per response}
\usage{
compute_R2(
  Y_list_test,
  X_list_test,
  indices_list_test,
  Y_list_train,
  indices_list_train,
  Beta
)
}
\arguments{
\item{Beta}{}
}
\value{

}
\description{
Compute R2 per response
}
