% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_SSE}
\alias{compute_SSE}
\title{Compute SSE per response}
\usage{
compute_SSE(Y_list, X_list, indices_list, Beta)
}
\arguments{
\item{Beta}{}
}
\value{

}
\description{
Compute SSE per response
}
