% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{compute_SST}
\alias{compute_SST}
\title{Compute SST per response}
\usage{
compute_SST(Y_list_test, indices_list_test, Y_list_train, indices_list_train)
}
\arguments{
\item{indices_list_train}{}
}
\value{

}
\description{
Compute SST per response
}
