% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{compute_tuning_performance}
\alias{compute_tuning_performance}
\title{Compute tuning performance on training/validation data}
\usage{
compute_tuning_performance(
  fit,
  Y_list,
  X_list,
  indices_list,
  Y_list_train,
  indices_list_train
)
}
\arguments{
\item{indices_list_train}{}
}
\value{

}
\description{
Compute tuning performance on training/validation data
}
