% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{compute_validation_performance}
\alias{compute_validation_performance}
\title{Compute performance on validation data}
\usage{
compute_validation_performance(
  fit,
  Y_list_validation,
  X_list_validation,
  indices_list_validation,
  Y_list_train,
  indices_list_train
)
}
\arguments{
\item{indices_list_train}{}
}
\value{

}
\description{
Compute performance on validation data
}
