% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{compute_weighted_avg_R2}
\alias{compute_weighted_avg_R2}
\title{Compute weighted average R2 across all responses with weights based on number of observations}
\usage{
compute_weighted_avg_R2(
  Y_list_test,
  X_list_test,
  indices_list_test,
  Y_list_train,
  indices_list_train,
  Beta
)
}
\arguments{
\item{Beta}{}
}
\value{

}
\description{
Compute weighted average R2 across all responses with weights based on number of observations
}
