% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compute_Y_list}
\alias{compute_Y_list}
\title{Compute list of Y matrices given X, Beta, L, E}
\usage{
compute_Y_list(X_list, Beta, L_list, E_list)
}
\arguments{
\item{E_list}{}
}
\value{

}
\description{
Compute list of Y matrices given X, Beta, L, E
}
