% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_utils.R
\name{construct_D_list}
\alias{construct_D_list}
\title{Construct D_{(k)} matrices based on observed responses}
\usage{
construct_D_list(Y_list)
}
\arguments{
\item{observed_response_list}{}
}
\value{

}
\description{
Construct D_{(k)} matrices based on observed responses
}
