% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{evaluate_parameters}
\alias{evaluate_parameters}
\title{Evaluate parameters}
\usage{
evaluate_parameters(parameters)
}
\arguments{
\item{parameters}{}
}
\value{

}
\description{
Evaluate parameters
}
