% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{fit_glmnet}
\alias{fit_glmnet}
\title{Fit glmnet model on concatenated data}
\usage{
fit_glmnet(
  Y_list,
  X_list,
  indices_list,
  Y_list_validation,
  X_list_validation,
  indices_list_validation,
  standardize = TRUE,
  n_lambda = 20,
  lambda_min_ratio = 0.001
)
}
\arguments{
\item{D_list_validation}{}
}
\value{

}
\description{
Fit glmnet model on concatenated data
}
