% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{generate_simulation_data}
\alias{generate_simulation_data}
\title{Generate data for simulation study}
\usage{
generate_simulation_data(
  R2,
  r,
  theta,
  sigma_2,
  sparsity,
  p,
  q,
  O_k_train,
  n_k_train,
  O_k_validation,
  n_k_validation,
  O_k_test,
  n_k_test,
  replicate
)
}
\arguments{
\item{n_k_test}{}
}
\value{

}
\description{
Generate data for simulation study
}
