% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLORS.R
\name{MultiLORS}
\alias{MultiLORS}
\title{Fit LORS model}
\usage{
MultiLORS(
  Y_list,
  X_list,
  D_list,
  Y_list_validation = NULL,
  X_list_validation = NULL,
  D_list_validation = NULL,
  standardize = TRUE,
  n_lambda = 20,
  n_gamma = 20,
  lambda_min_ratio = 0.001,
  gamma_min_ratio = 0.001,
  n_iter = 1000,
  tolerance = 1e-06,
  early_stopping = TRUE,
  line_search = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Fit LORS model
}
