% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_actual_vs_predicted}
\alias{plot_actual_vs_predicted}
\title{Plot actual versus predicted values for each response variable}
\usage{
plot_actual_vs_predicted(
  Y_list,
  X_list,
  indices_list,
  Beta,
  annotations = NULL
)
}
\arguments{
\item{annotations}{}
}
\value{

}
\description{
Plot actual versus predicted values for each response variable
}
