% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_Y_and_indices_test}
\alias{prepare_Y_and_indices_test}
\title{Make ordering of observed responses consistent and construct vectors indicating indices of all responses observed, based on responses available in training data}
\usage{
prepare_Y_and_indices_test(Y_list_test, indices_list_train)
}
\arguments{
\item{Y_list_test}{}
}
\value{

}
\description{
Make ordering of observed responses consistent and construct vectors indicating indices of all responses observed, based on responses available in training data
}
