% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_Y_and_indices_train}
\alias{prepare_Y_and_indices_train}
\title{Make ordering of observed responses consistent and construct vectors indicating indices of all responses observed}
\usage{
prepare_Y_and_indices_train(Y_list_train)
}
\arguments{
\item{Y_list_train}{}
}
\value{

}
\description{
Make ordering of observed responses consistent and construct vectors indicating indices of all responses observed
}
