% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{refit_glmnet}
\alias{refit_glmnet}
\title{Refit glmnet on nonzero coefficients using OLS}
\usage{
refit_glmnet(
  fit,
  Y_list,
  X_list,
  indices_list,
  Y_list_validation = NULL,
  X_list_validation = NULL,
  indices_list_validation = NULL
)
}
\arguments{
\item{indices_list_validation}{}
}
\value{

}
\description{
Refit glmnet on nonzero coefficients using OLS
}
