% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLORS.R
\name{refit_MultiLORS}
\alias{refit_MultiLORS}
\title{Refit MultiLORS on nonzero coefficients and corrected responses using OLS}
\usage{
refit_MultiLORS(
  fit,
  Y_list,
  X_list,
  indices_list,
  Y_list_validation = NULL,
  X_list_validation = NULL,
  indices_list_validation = NULL
)
}
\arguments{
\item{indices_list_validation}{}
}
\value{

}
\description{
Refit MultiLORS on nonzero coefficients and corrected responses using OLS
}
