% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_Beta}
\alias{simulate_Beta}
\title{simulate row-sparse Beta matrix}
\usage{
simulate_Beta(p, q, sparsity, R2 = NULL, X = NULL, L = NULL, E = NULL)
}
\arguments{
\item{q}{}
}
\value{

}
\description{
simulate row-sparse Beta matrix
}
