% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_E}
\alias{simulate_E}
\title{simulate E error matrix with specified covariance matrix}
\usage{
simulate_E(n, Sigma)
}
\arguments{
\item{Sigma}{}
}
\value{

}
\description{
simulate E error matrix with specified covariance matrix
}
