% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_E_list}
\alias{simulate_E_list}
\title{simulate list of E error matrices with specified covariance matrix}
\usage{
simulate_E_list(n_k, Sigma)
}
\arguments{
\item{Sigma}{}
}
\value{

}
\description{
simulate list of E error matrices with specified covariance matrix
}
