% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_L_list}
\alias{simulate_L_list}
\title{simulate list of L batch effect matrices}
\usage{
simulate_L_list(n_k, q, theta)
}
\arguments{
\item{variance}{}
}
\value{

}
\description{
simulate list of L batch effect matrices
}
