% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{subset_simulated_data}
\alias{subset_simulated_data}
\title{Subset simulated data to have different observed responses in each dataset}
\usage{
subset_simulated_data(Y_list, L_list, E_list, O_k)
}
\arguments{
\item{O_k}{}
}
\value{

}
\description{
Subset simulated data to have different observed responses in each dataset
}
