% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{which_max}
\alias{which_max}
\title{Determine indices of maximum element in matrix}
\usage{
which_max(mat)
}
\arguments{
\item{mat}{}
}
\value{

}
\description{
Determine indices of maximum element in matrix
}
