% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{which_min}
\alias{which_min}
\title{Determine indices of minimum element in matrix}
\usage{
which_min(mat)
}
\arguments{
\item{mat}{}
}
\value{

}
\description{
Determine indices of minimum element in matrix
}
