% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLORS.R
\name{MultiLORS}
\alias{MultiLORS}
\title{Fit MultiLORS model}
\usage{
MultiLORS(
  Y_list,
  X_list,
  indices_list,
  Y_list_validation = NULL,
  X_list_validation = NULL,
  indices_list_validation = NULL,
  standardize_Y = FALSE,
  standardize_X = TRUE,
  gamma_method = 1,
  n_lambda = 20,
  n_gamma = 20,
  lambda_min_ratio = 0.001,
  gamma_min_ratio = 0.001,
  n_iter = 1000,
  tolerance = 1e-06,
  early_stopping = TRUE,
  verbose = 0,
  return_L = TRUE,
  n_cores = 1
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Fit MultiLORS model
}
